Use BooksAPI to render books with correct content
Since each Book knows it's ID, we can use the API to get the information
needed to render the Book with the correct content.

Although this may seem like an inefficent way to do this, as we've
already queried `BooksAPI.getAll`, HTTP caching makes this operation
relatively inexpensive in the long term. It would be nice to see if you
could add a local caching layer, but we'll see if there's time for that
later.